Implemented stack functions

+++ .ccls-cache/@home@runner@LinkedList-Functions/queueFunctions.h
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 
 typedef struct doubleNode {
-  int key;
+  double key;
   struct doubleNode* next;
 } doubleNode;
